requirements.txt
Flask==2.2.3
Pillow==9.5.0
gTTS==2.2.3
transformers==4.29.2
gunicorn==20.1.0
torch==1.11.0


